requirements.txt
elasticsearch==8.11.0
langchain==0.1.0
langchain-community==0.0.10
anthropic==0.25.0
fastapi==0.109.0
uvicorn==0.25.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
aiohttp==3.9.1
ipython==8.18.0
jupyter==1.0.0
jsonrpc-websocket==3.1.4